refactor DocumentFilterDescriptorsProvider(s)
https://github.com/metasfresh/metasfresh/issues/5101
