+++ src/main/java/de/metas/ui/web/document/filter/json/JSONDocumentFilter.java
@@ -14,9 +14,9 @@ import com.google.common.collect.Maps;
 
 import de.metas.ui.web.document.filter.DocumentFilter;
 import de.metas.ui.web.document.filter.DocumentFilterDescriptor;
-import de.metas.ui.web.document.filter.DocumentFilterDescriptorsProvider;
 import de.metas.ui.web.document.filter.DocumentFilterParam;
 import de.metas.ui.web.document.filter.DocumentFilterParamDescriptor;
+import de.metas.ui.web.document.filter.provider.DocumentFilterDescriptorsProvider;
 import de.metas.util.Check;
 import de.metas.util.GuavaCollectors;
