+++ src/main/java/de/metas/ui/web/document/filter/provider/NullDocumentFilterDescriptorsProvider.java
@@ -1,10 +1,12 @@
-package de.metas.ui.web.document.filter;
+package de.metas.ui.web.document.filter.provider;
 
 import java.util.Collection;
 
 import javax.annotation.concurrent.Immutable;
 
 import com.google.common.collect.ImmutableList;
+
+import de.metas.ui.web.document.filter.DocumentFilterDescriptor;
 
 /*
  * #%L
@@ -19,11 +21,11 @@ import com.google.common.collect.ImmutableList;
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
@@ -32,15 +34,19 @@ import com.google.common.collect.ImmutableList;
 public final class NullDocumentFilterDescriptorsProvider implements DocumentFilterDescriptorsProvider
 {
 	public static final transient NullDocumentFilterDescriptorsProvider instance = new NullDocumentFilterDescriptorsProvider();
-	
+
 	public static final boolean isNull(final DocumentFilterDescriptorsProvider provider)
 	{
 		return provider == null || provider == instance;
 	}
-	
+
+	public static final boolean isNotNull(final DocumentFilterDescriptorsProvider provider)
+	{
+		return !isNull(provider);
+	}
+
 	private NullDocumentFilterDescriptorsProvider()
 	{
-		super();
 	}
 
 	@Override
