+++ src/main/java/de/metas/ui/web/document/filter/provider/standard/StandardDocumentFilterDescriptorsProviderFactory.java
@@ -1,4 +1,4 @@
-package de.metas.ui.web.document.filter;
+package de.metas.ui.web.document.filter.provider.standard;
 
 import java.util.Collection;
 import java.util.List;
@@ -8,23 +8,21 @@ import java.util.stream.Stream;
 import javax.annotation.Nullable;
 
 import org.adempiere.ad.element.api.AdTabId;
-import org.compiere.Adempiere;
-import org.elasticsearch.client.Client;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Ordering;
 
-import de.metas.elasticsearch.indexer.IESModelIndexer;
-import de.metas.elasticsearch.indexer.IESModelIndexersRegistry;
 import de.metas.i18n.IMsgBL;
 import de.metas.i18n.ITranslatableString;
+import de.metas.ui.web.document.filter.DocumentFilterDescriptor;
 import de.metas.ui.web.document.filter.DocumentFilterParam.Operator;
+import de.metas.ui.web.document.filter.DocumentFilterParamDescriptor;
+import de.metas.ui.web.document.filter.provider.DocumentFilterDescriptorsProvider;
+import de.metas.ui.web.document.filter.provider.ImmutableDocumentFilterDescriptorsProvider;
 import de.metas.ui.web.window.descriptor.DocumentFieldDefaultFilterDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentFieldDescriptor.Characteristic;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
-import de.metas.ui.web.window.descriptor.FullTextSearchFilterContext;
-import de.metas.ui.web.window.descriptor.FullTextSearchSqlDocumentFilterConverter;
 import de.metas.ui.web.window.descriptor.LookupDescriptor;
 import de.metas.util.Services;
 
@@ -32,66 +30,47 @@ import de.metas.util.Services;
  * #%L
  * metasfresh-webui-api
  * %%
- * Copyright (C) 2016 metas GmbH
+ * Copyright (C) 2019 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-/**
- * {@link DocumentFilterDescriptorsProvider}s factory.
- *
- * @author metas-dev <dev@metasfresh.com>
- *
- */
-public final class DocumentFilterDescriptorsProviderFactory
+public class StandardDocumentFilterDescriptorsProviderFactory
 {
+	public static final transient StandardDocumentFilterDescriptorsProviderFactory instance = new StandardDocumentFilterDescriptorsProviderFactory();
+
 	private static final String FILTER_ID_Default = "default";
 
 	private static final String FILTER_ID_DefaultDate = "default-date";
 	private static final String MSG_DefaultFilterName = "default";
 
-	private static final String MSG_FULL_TEXT_SEARCH_CAPTION = "Search";
-
-	public static final transient DocumentFilterDescriptorsProviderFactory instance = new DocumentFilterDescriptorsProviderFactory();
-
 	// services
 	private final transient IMsgBL msgBL = Services.get(IMsgBL.class);
 
-	private DocumentFilterDescriptorsProviderFactory()
+	private StandardDocumentFilterDescriptorsProviderFactory()
 	{
-		super();
-	}
-
-	public DocumentFilterDescriptorsProvider createFiltersProvider(
-			final AdTabId adTabId, 
-			@Nullable final String tableName, 
-			final Collection<DocumentFieldDescriptor> fields)
-	{
-		return CompositeDocumentFilterDescriptorsProvider.compose(
-				createUserQueryDocumentFilterDescriptorsProvider(adTabId, tableName, fields),
-				createFiltersProvider_Defaults(fields),
-				createFullTextSeachFilterDescriptorProvider(tableName));
 	}
 
 	/**
 	 * Creates standard filters, i.e. from document fields which are flagged with {@link Characteristic#AllowFiltering}.
-	 *
-	 * @param fields
 	 */
-	private ImmutableDocumentFilterDescriptorsProvider createFiltersProvider_Defaults(final Collection<DocumentFieldDescriptor> fields)
+	public DocumentFilterDescriptorsProvider createFiltersProvider(
+			@Nullable final AdTabId adTabId_NOTUSED,
+			@Nullable final String tableName_NOTUSED,
+			final Collection<DocumentFieldDescriptor> fields)
 	{
 		final DocumentFilterDescriptor.Builder defaultFilter = DocumentFilterDescriptor.builder()
 				.setFilterId(FILTER_ID_Default)
@@ -126,7 +105,7 @@ public final class DocumentFilterDescriptorsProviderFactory
 				.collect(ImmutableDocumentFilterDescriptorsProvider.collector());
 	}
 
-	private final DocumentFilterParamDescriptor.Builder createFilterParam(final DocumentFieldDescriptor field)
+	private static final DocumentFilterParamDescriptor.Builder createFilterParam(final DocumentFieldDescriptor field)
 	{
 		final ITranslatableString displayName = field.getCaption();
 		final String fieldName = field.getFieldName();
@@ -160,7 +139,7 @@ public final class DocumentFilterDescriptorsProviderFactory
 				.setAutoFilterInitialValue(filteringInfo.getAutoFilterInitialValue());
 	}
 
-	private DocumentFieldWidgetType extractFilterWidgetType(final DocumentFieldDescriptor field)
+	private static DocumentFieldWidgetType extractFilterWidgetType(final DocumentFieldDescriptor field)
 	{
 		final DocumentFieldWidgetType widgetType = field.getWidgetType();
 		if (widgetType == DocumentFieldWidgetType.DateTime
@@ -168,68 +147,10 @@ public final class DocumentFilterDescriptorsProviderFactory
 		{
 			return DocumentFieldWidgetType.Date;
 		}
-
-		return widgetType;
-
-	}
-
-	private static DocumentFilterDescriptorsProvider createUserQueryDocumentFilterDescriptorsProvider(
-			final AdTabId adTabId,
-			@Nullable final String tableName,
-			final Collection<DocumentFieldDescriptor> fields)
-	{
-		if (tableName != null && adTabId != null)
-		{
-			return new UserQueryDocumentFilterDescriptorsProvider(adTabId, tableName, fields);
-		}
 		else
 		{
-			return NullDocumentFilterDescriptorsProvider.instance;
+			return widgetType;
 		}
 	}
 
-	private DocumentFilterDescriptorsProvider createFullTextSeachFilterDescriptorProvider(@Nullable final String modelTableName)
-	{
-		if (modelTableName == null)
-		{
-			return NullDocumentFilterDescriptorsProvider.instance;
-		}
-
-		final IESModelIndexersRegistry esModelIndexersRegistry = Services.get(IESModelIndexersRegistry.class);
-		final IESModelIndexer modelIndexer = esModelIndexersRegistry.getFullTextSearchModelIndexer(modelTableName)
-				.orElse(null);
-		if (modelIndexer == null)
-		{
-			return NullDocumentFilterDescriptorsProvider.instance;
-		}
-
-		final ITranslatableString caption = msgBL.getTranslatableMsgText(MSG_FULL_TEXT_SEARCH_CAPTION);
-		final FullTextSearchFilterContext context = createFullTextSearchFilterContext(modelIndexer);
-
-		final DocumentFilterDescriptor filterDescriptor = DocumentFilterDescriptor.builder()
-				.setFilterId(FullTextSearchSqlDocumentFilterConverter.FILTER_ID)
-				.setDisplayName(caption)
-				.setFrequentUsed(true)
-				.setInlineRenderMode(DocumentFilterInlineRenderMode.INLINE_PARAMETERS)
-				.addParameter(DocumentFilterParamDescriptor.builder()
-						.setFieldName(FullTextSearchSqlDocumentFilterConverter.PARAM_SearchText)
-						.setDisplayName(caption)
-						.setWidgetType(DocumentFieldWidgetType.Text))
-				.addInternalParameter(DocumentFilterParam.ofNameEqualsValue(FullTextSearchSqlDocumentFilterConverter.PARAM_Context, context))
-				.build();
-
-		return ImmutableDocumentFilterDescriptorsProvider.of(filterDescriptor);
-	}
-
-	private FullTextSearchFilterContext createFullTextSearchFilterContext(final IESModelIndexer modelIndexer)
-	{
-		final Client elasticsearchClient = Adempiere.getBean(org.elasticsearch.client.Client.class);
-
-		return FullTextSearchFilterContext.builder()
-				.elasticsearchClient(elasticsearchClient)
-				.modelTableName(modelIndexer.getModelTableName())
-				.esIndexName(modelIndexer.getIndexName())
-				.esSearchFieldNames(modelIndexer.getFullTextSearchFieldNames())
-				.build();
-	}
 }
