+++ src/main/java/de/metas/ui/web/material/cockpit/stockdetails/StockDetailsView.java
@@ -1,7 +1,7 @@
 package de.metas.ui.web.material.cockpit.stockdetails;
 
 import de.metas.handlingunits.model.I_M_HU_Stock_Detail_V;
-import de.metas.ui.web.document.filter.DocumentFilterDescriptorsProvider;
+import de.metas.ui.web.document.filter.provider.DocumentFilterDescriptorsProvider;
 import de.metas.ui.web.view.AbstractCustomView;
 import de.metas.ui.web.view.IView;
 import de.metas.ui.web.view.ViewId;
