+++ src/main/java/de/metas/ui/web/window/descriptor/sql/SqlEntityBinding.java
@@ -4,7 +4,7 @@ import java.util.regex.Pattern;
 
 import org.adempiere.ad.expression.api.IStringExpression;
 
-import de.metas.ui.web.document.filter.DocumentFilterDescriptorsProvider;
+import de.metas.ui.web.document.filter.provider.DocumentFilterDescriptorsProvider;
 import de.metas.ui.web.document.filter.sql.SqlDocumentFilterConverterDecorator;
 import de.metas.ui.web.document.filter.sql.SqlDocumentFilterConverters;
 import de.metas.ui.web.document.filter.sql.SqlDocumentFilterConvertersList;
